+++ scripts/winesapos-uninstall.sh
@@ -13,10 +13,17 @@ crudini --del /etc/pacman.conf jupiter-rel
 crudini --del /etc/pacman.conf holo-rel
 pacman -S -y
 
+for svc in \
+  winesapos-plasmalogin-health-check \
+  winesapos-sddm-health-check;
+    do systemctl disable "${svc}"
+done
+
 rm -r -f \
   /etc/modprobe.d/winesapos-* \
   /etc/os-release-winesapos \
   /etc/sddm.conf.d/winesapos.conf \
+  /etc/plasmalogin.conf.d/winesapos.conf \
   /etc/sysctl.d/*winesapos* \
   /etc/systemd/system/pacman-mirrors.service \
   /etc/systemd/system/snapper-cleanup-hourly.timer \
@@ -31,8 +38,10 @@ rm -r -f \
   /usr/lib/sysctl.d/*winesapos* \
   /usr/local/bin/winesapos-dual-boot.sh \
   /usr/local/bin/winesapos-resize-root-file-system.sh \
+  /usr/local/bin/winesapos-*-health-check.sh \
   /usr/share/libalpm/hooks/winesapos-*.hook \
   /usr/share/sddm/faces/winesap.face.icon \
+  /var/lib/AccountsService/icons/winesap \
   /var/winesapos/
 
 systemctl daemon-reload